# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: "0.2.0"
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List groups the current user belongs to
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a new group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Created group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{groupId}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get a group's detail including members
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Group detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupDetail"
        "404":
          description: Not found
    delete:
      operationId: deleteGroup
      tags: [groups]
      summary: Delete an entire group
      description: >
        Only the group's creator can do this. Deletes the group and
        everything in it (members, messages, encryption keys) via cascading
        foreign keys, and notifies any currently-connected members over
        WebSocket so their clients can leave the chat immediately.
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "403":
          description: Not the group's creator
        "404":
          description: Not found
  /groups/{groupId}/avatar:
    put:
      operationId: setGroupAvatar
      tags: [groups]
      summary: Set or clear a group's photo
      description: >
        Body is a small base64 data URI — the client resizes to a thumbnail
        before uploading. Not E2E-encrypted (same tradeoff as a user's
        profile photo). Pass null to remove the current photo. Any member
        may change it.
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetGroupAvatarInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetGroupAvatarInput"
        "400":
          description: Image too large or not a data URI
        "404":
          description: Not found
  /groups/{groupId}/members:
    post:
      operationId: addGroupMember
      tags: [groups]
      summary: Add a member to a group by email
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupMemberInput"
      responses:
        "201":
          description: Added member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupMember"
        "404":
          description: User with that email not found
  /groups/{groupId}/key:
    get:
      operationId: getMyGroupKey
      tags: [groups]
      summary: Get the current user's wrapped end-to-end encryption key for a group
      description: >
        Returns the group's symmetric encryption key, wrapped with the
        current user's public key. Returns wrappedKey null if it has not
        been shared with this user yet.
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Wrapped group key, if shared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupKeyResponse"
        "404":
          description: Not found
  /groups/{groupId}/keys:
    post:
      operationId: setGroupKey
      tags: [groups]
      summary: Share a wrapped copy of the group's encryption key with a member
      description: >
        Stores a copy of the group's symmetric key, wrapped with the target
        member's public key. Called by any member who already holds the
        decrypted group key (e.g. the creator, or when re-sharing with a
        newly-invited member).
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupKeyInput"
      responses:
        "201":
          description: Stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupKeyResponse"
        "404":
          description: Not found
  /groups/{groupId}/keys/request:
    post:
      operationId: requestGroupKeyAccess
      tags: [groups]
      summary: Ask other connected members to re-share the group key with me
      description: >
        Used when this browser's copy of the group key is missing (e.g. its
        private key was lost when the browser's storage was cleared). Pings
        any other member currently connected to this group over WebSocket;
        if their client already holds the decrypted group key, it re-wraps
        and re-shares it for this user automatically. Best-effort — if no
        other member is currently viewing this group, nothing happens until
        one of them opens it (or requests again later).
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Request broadcast (delivery not guaranteed)
        "404":
          description: Not found
  /groups/{groupId}/read:
    put:
      operationId: markGroupRead
      tags: [groups]
      summary: Mark a group as read up to now, for the current user
      description: >
        Sets the current user's last-read timestamp for this group to now.
        Powers unread badges (client compares each message's createdAt
        against this timestamp) and "Seen" receipts (the sender compares
        their last message's createdAt against every other member's
        last-read timestamp). Broadcasts a WS "read" event so anyone
        currently viewing the group updates seen-status live.
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated read receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadReceipt"
        "404":
          description: Not found
  /groups/{groupId}/pin:
    put:
      operationId: setGroupPinned
      tags: [groups]
      summary: Pin or unpin a group in the current user's own chat list
      description: >
        Purely personal — this only affects how the group sorts in the
        caller's own list, not for any other member.
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPinnedInput"
      responses:
        "200":
          description: Updated pin state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PinnedResult"
        "404":
          description: Not found
  /groups/{groupId}/mute:
    put:
      operationId: setGroupMuted
      tags: [groups]
      summary: Mute or unmute a group for the current user
      description: >
        Purely personal — silences notifications for this group for the
        caller only. Never affects the caller's ability to send or receive
        messages, or any other member's settings.
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetMutedInput"
      responses:
        "200":
          description: Updated mute state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MutedResult"
        "404":
          description: Not found
  /groups/{groupId}/calls/join:
    post:
      operationId: joinGroupCall
      tags: [groups]
      summary: Join (starting if needed) this group's call
      description: >
        Group calls are a shared room, not a 1:1 ring — unlike DM calls,
        there's no per-person "missed" tracking, no ring, and no push. If a
        call is already active for this group, joins that one (no new log
        entry); otherwise starts a new one. A compact summary ("Video call
        · 12m") is written into the group's messages once the last
        participant leaves (see the /leave endpoint).
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinGroupCallInput"
      responses:
        "200":
          description: Joined (or started) the call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinGroupCallResult"
        "404":
          description: Not found
...
components:
  schemas:
...
    ReadReceipt:
      type: object
      required: [lastReadAt]
      properties:
        lastReadAt:
          type: string
    SetGroupAvatarInput:
      type: object
      required: [avatarUrl]
      properties:
        avatarUrl:
          type: ["string", "null"]
...
    Group:
      type: object
      required:
        [
          id,
          name,
          createdBy,
          createdAt,
          memberCount,
          unreadCount,
          isPinned,
          isMuted,
        ]
      properties:
        id:
          type: string
        name:
          type: string
        createdBy:
          type: string
        createdAt:
          type: string
        avatarUrl:
          type: ["string", "null"]
        memberCount:
          type: integer
        lastMessageAt:
          type: ["string", "null"]
        lastMessagePreview:
          type: ["string", "null"]
        myLastReadAt:
          type: ["string", "null"]
        unreadCount:
          type: integer
          description: Messages from other members created after myLastReadAt.
        isPinned:
          type: boolean
          description: >
            Whether the current user has pinned this group to the top of
            their own chat list. Purely personal — pinning is per-member,
            not a group-wide setting.
        isMuted:
          type: boolean
          description: >
            Whether the current user has muted notifications for this
            group. Purely personal, like isPinned — never affects sending
            or receiving messages.
    GroupInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
...
    GroupMemberInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
...